Fix Consumer naming
