+++ consumers/docs.go
@@ -1,6 +1,6 @@
 // Copyright (c) Mainflux
 // SPDX-License-Identifier: Apache-2.0
 
-// Package writers contain the domain concept definitions needed to
+// Package consumers contain the domain concept definitions needed to
 // support Mainflux writer services functionality.
 package consumers
