+++ consumers/writers/cassandra/consumer_test.go
