+++ consumers/writers/influxdb/consumer_test.go
