+++ consumers/writers/postgres/consumer.go
@@ -31,14 +31,14 @@ var (
 	errNoTable        = errors.New("relation does not exist")
 )
 
-var _ consumers.MessageConsumer = (*postgresRepo)(nil)
+var _ consumers.Consumer = (*postgresRepo)(nil)
 
 type postgresRepo struct {
 	db *sqlx.DB
 }
 
 // New returns new PostgreSQL writer.
-func New(db *sqlx.DB) consumers.MessageConsumer {
+func New(db *sqlx.DB) consumers.Consumer {
 	return &postgresRepo{db: db}
 }
