--- consumers/writers/writer.go
@@ -1,83 +1,0 @@
-// Copyright (c) Mainflux
-// SPDX-License-Identifier: Apache-2.0
-
-package writers
-
-import (
-	"fmt"
-	"io/ioutil"
-
-	"github.com/BurntSushi/toml"
-	"github.com/mainflux/mainflux/consumers"
-	"github.com/mainflux/mainflux/logger"
-	"github.com/mainflux/mainflux/pkg/errors"
-	"github.com/mainflux/mainflux/pkg/messaging"
-	pubsub "github.com/mainflux/mainflux/pkg/messaging/nats"
-	"github.com/mainflux/mainflux/pkg/transformers"
-)
-
-var (
-	errOpenConfFile      = errors.New("unable to open configuration file")
-	errParseConfFile     = errors.New("unable to parse configuration file")
-	errMessageConversion = errors.New("error conversing transformed messages")
-)
-
-type writer struct {
-	consumer    consumers.MessageConsumer
-	transformer transformers.Transformer
-	logger      logger.Logger
-}
-
-// Start method starts consuming messages received from NATS.
-// This method transforms messages to SenML format before
-// using MessageRepository to store them.
-func Start(sub messaging.Subscriber, consumer consumers.MessageConsumer, transformer transformers.Transformer, subjectsCfgPath string, logger logger.Logger) error {
-	w := writer{
-		consumer:    consumer,
-		transformer: transformer,
-		logger:      logger,
-	}
-
-	subjects, err := loadSubjectsConfig(subjectsCfgPath)
-	if err != nil {
-		logger.Warn(fmt.Sprintf("Failed to load subjects: %s", err))
-	}
-
-	for _, subject := range subjects {
-		if err := sub.Subscribe(subject, w.handler); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (c *writer) handler(msg messaging.Message) error {
-	t, err := c.transformer.Transform(msg)
-	if err != nil {
-		return err
-	}
-
-	return c.consumer.Consume(t)
-}
-
-type filterConfig struct {
-	Filter []string `toml:"filter"`
-}
-
-type subjectsConfig struct {
-	Subjects filterConfig `toml:"subjects"`
-}
-
-func loadSubjectsConfig(subjectsConfigPath string) ([]string, error) {
-	data, err := ioutil.ReadFile(subjectsConfigPath)
-	if err != nil {
-		return []string{pubsub.SubjectAllChannels}, errors.Wrap(errOpenConfFile, err)
-	}
-
-	var subjectsCfg subjectsConfig
-	if err := toml.Unmarshal(data, &subjectsCfg); err != nil {
-		return []string{pubsub.SubjectAllChannels}, errors.Wrap(errParseConfFile, err)
-	}
-
-	return subjectsCfg.Subjects.Filter, nil
-}
